requirements.txt
backports.weakref==1.0rc1
bleach==1.5.0
boto==2.48.0
boto3==1.7.19
botocore==1.10.19
bz2file==0.98
certifi==2017.11.5
cffi==1.11.5
chardet==3.0.4
click==6.7
cycler==0.10.0
definitions==0.2.0
docutils==0.14
enum34==1.1.6
Flask==1.0.2
gensim==3.4.0
gevent==1.3.0
greenlet==0.4.13
h5py==2.7.0
html5lib==0.9999999
idna==2.6
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
Keras==2.0.6
lxml==4.1.1
Mako==1.0.6
Markdown==2.6.10
MarkupSafe==0.23
matplotlib==2.1.2
nltk==3.2.5
nose==1.3.7
numpy==1.14.0
pandas==0.22.0
Pillow==5.1.0
protobuf==3.5.1
pycparser==2.18
pygpu==0.6.8
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.3
PyYAML==3.12
requests==2.18.4
s3transfer==0.1.13
scikit-learn==0.19.1
scipy==0.19.1
sets==0.3.2
six==1.11.0
sklearn==0.0
smart-open==1.5.7
tensorflow==1.4.0
tensorflow-hub==0.1.0
tensorflow-tensorboard==0.4.0rc3
Theano==0.9.0
tqdm==4.23.3
urllib3==1.22
Werkzeug==0.14.1
